Progress syncing with latest TC, including waypoint, map, group/loot & instance systems.

+++ src/server/game/Server/Packets/CalendarPackets.cpp
@@ -429,7 +429,7 @@ WorldPacket const* WorldPackets::Calendar::CalendarCommandResult::Write()
 WorldPacket const* WorldPackets::Calendar::CalendarRaidLockoutAdded::Write()
 {
     _worldPacket << uint64(InstanceID);
-    _worldPacket << uint32(ServerTime);
+    _worldPacket << ServerTime;
     _worldPacket << int32(MapID);
     _worldPacket << uint32(DifficultyID);
     _worldPacket << int32(TimeRemaining);
